+++ frontend/src/styles/ContactUs.css
@@ -1,10 +1,8 @@
 .portal-contact {
-  position: fixed;
-  right: clamp(1rem, 3vw, 2rem);
-  bottom: calc(clamp(1rem, 3vw, 2rem) + 4.2rem + 0.9rem);
-  z-index: 41;
-  width: 4.2rem;
-  height: 4.2rem;
+  position: relative;
+  z-index: 1;
+  width: var(--portal-floating-button-size, 4.2rem);
+  height: var(--portal-floating-button-size, 4.2rem);
   pointer-events: none;
 }
 
@@ -14,8 +12,8 @@
   bottom: 0;
   display: grid;
   place-items: center;
-  width: 4.2rem;
-  height: 4.2rem;
+  width: var(--portal-floating-button-size, 4.2rem);
+  height: var(--portal-floating-button-size, 4.2rem);
   border: 1px solid rgba(250, 204, 21, 0.28);
   border-radius: 9999px;
   color: #fff7ed;
@@ -207,13 +205,6 @@
 }
 
 @media (max-width: 640px) {
-  .portal-contact {
-    right: 0.85rem;
-    bottom: calc(max(0.85rem, env(safe-area-inset-bottom)) + 3.65rem + 0.75rem);
-    width: 3.65rem;
-    height: 3.65rem;
-  }
-
   .portal-contact__panel {
     right: calc(100% + 0.75rem);
     width: min(22rem, calc(100vw - 5.5rem));
@@ -237,11 +228,6 @@
   .portal-contact__form {
     gap: 0.9rem;
     padding: 0.85rem 0.85rem 1rem;
-  }
-
-  .portal-contact__button {
-    width: 3.65rem;
-    height: 3.65rem;
   }
 }
